v3.5

+++ src/Servebolt/CachePurge/ThirdPartyFunctions.php
@@ -1,19 +1,23 @@
 <?php
 
+use Servebolt\Optimizer\CachePurge\CachePurge;
 use Servebolt\Optimizer\CachePurge\WordPressCachePurge\WordPressCachePurge;
 use Servebolt\Optimizer\Exceptions\ApiError;
 
 /**
  * Purge cache for URL.
  *
- * @param string $url
- * @param bool $return_wp_error_object
+ * @param string $url The URL to purge cache for.
+ * @param bool $return_wp_error_object Whether to return WP_Error-object on error.
  * @return bool|WP_Error
  */
-function sb_purge_url(string $url, bool $return_wp_error_object = false)
+function sb_purge_url($url, bool $return_wp_error_object = false)
 {
+    if (!CachePurge::driverSupportsUrlCachePurge()) {
+        return false; // Current cache purge driver does not support purging of URLs.
+    }
     try {
-        return WordPressCachePurge::purgeByUrl($url);
+        return WordPressCachePurge::purgeByUrl((string) $url);
     } catch (ApiError $e) {
         if ($return_wp_error_object) {
             // TODO: Return WP Error object
@@ -30,12 +34,15 @@ function sb_purge_url(string $url, bool $return_wp_error_object = false)
 /**
  * Purge cache for post.
  *
- * @param $post_id
- * @param bool $return_wp_error_object
+ * @param int $post_id The ID of the post to purge cache for.
+ * @param bool $return_wp_error_object Whether to return WP_Error-object on error.
  * @return bool|WP_Error
  */
 function sb_purge_post_cache($post_id, bool $return_wp_error_object = false)
 {
+    if (!CachePurge::driverSupportsItemCachePurge()) {
+        return false; // Current cache purge driver does not support purging of URLs/posts.
+    }
     try {
         return WordPressCachePurge::purgeByPostId(
             (int) $post_id
@@ -59,17 +66,20 @@ function sb_purge_post_cache($post_id, bool $return_wp_error_object = false)
 /**
  * Purge cache for term.
  *
- * @param $term_id
- * @param $taxonomy_slug
- * @param bool $return_wp_error_object
+ * @param int $term_id The ID of the term to purge cache for.
+ * @param string $taxonomy_slug The slug of the taxonomy of the term.
+ * @param bool $return_wp_error_object Whether to return WP_Error-object on error.
  * @return bool|WP_Error
  */
 function sb_purge_term_cache($term_id, $taxonomy_slug, bool $return_wp_error_object = false)
 {
+    if (!CachePurge::driverSupportsItemCachePurge()) {
+        return false; // Current cache purge driver does not support purging of URLs/terms.
+    }
     try {
         return WordPressCachePurge::purgeByTermId(
             (int) $term_id,
-            $taxonomy_slug
+            (string) $taxonomy_slug
         );
     } catch (ApiError $e) {
         // TODO: Handle API error message
@@ -90,11 +100,14 @@ function sb_purge_term_cache($term_id, $taxonomy_slug, bool $return_wp_error_obj
 /**
  * Purge all cache.
  *
- * @param bool $return_wp_error_object
+ * @param bool $return_wp_error_object Whether to return WP_Error-object on error.
  * @return bool|WP_Error
  */
 function sb_purge_all(bool $return_wp_error_object = false)
 {
+    if (!CachePurge::driverSupportsCachePurgeAll()) {
+        return false; // Current cache purge driver does not support purge all.
+    }
     try {
         return WordPressCachePurge::purgeAll();
     } catch (ApiError $e) {
@@ -115,16 +128,20 @@ function sb_purge_all(bool $return_wp_error_object = false)
 /**
  * Purge all cache by blog Id (multisite only).
  *
- * @param $blog_id
- * @param bool $return_wp_error_object
+ * @param int $blog_id The ID of the blog to purge all cache for.
+ * @param bool $return_wp_error_object Whether to return WP_Error-object on error.
  * @return false|WP_Error
  */
 function sb_purge_all_by_blog_id($blog_id, bool $return_wp_error_object = false)
 {
+    if (!is_multisite()) {
+        return false;
+    }
+    if (!CachePurge::driverSupportsCachePurgeAll()) {
+        return false; // Current cache purge driver does not support purge all.
+    }
     try {
-        return WordPressCachePurge::purgeAllByBlogId(
-            (int) $blog_id
-        );
+        return WordPressCachePurge::purgeAllByBlogId((int) $blog_id);
     } catch (ApiError $e) {
         // TODO: Handle API error message
         if ($return_wp_error_object) {
@@ -143,11 +160,18 @@ function sb_purge_all_by_blog_id($blog_id, bool $return_wp_error_object = false)
 /**
  * Purge all cache in network (multisite only).
  *
- * @param bool $return_wp_error_object
+ * @param bool $return_wp_error_object Whether to return WP_Error-object on error.
  * @return bool|WP_Error
  */
 function sb_purge_all_network(bool $return_wp_error_object = false)
 {
+    if (!is_multisite()) {
+        return false;
+    }
+    // TODO: Add multisite support for this check
+    if (!CachePurge::driverSupportsCachePurgeAll()) {
+        return false; // Current cache purge driver does not support purge all.
+    }
     try {
         return WordPressCachePurge::purgeAllNetwork();
     } catch (ApiError $e) {
